+++ R/Args.R
@@ -210,9 +210,11 @@ fetchRdDB <- function(filebase, key = NULL) {
 #' # [1] 24
 #' @export
 selectFromList <- function(x, select) {
-  return(sapply(x, function(x) {
-    x[names(x)[names(x) %in% select]]
-  }, simplify = FALSE))
+  selectFun <- function(y) {
+    fields <- names(y)[names(y) %in% select]
+    return(sapply(fields, function(z) y[[z]], simplify = FALSE))
+  }
+  return(sapply(x, selectFun, simplify = FALSE))
 }
 
 #' Exclude variables from a list of objects of the same type